+++ sdk/search/Azure.Search.Documents/tests/Utilities/SearchTestBase.cs
@@ -20,7 +20,7 @@ namespace Azure.Search.Documents.Tests
     /// Base class for Search unit tests that adds shared infrastructure on top
     /// of the Azure.Core testing framework.
     /// </summary>
-    [ClientTestFixture(SearchClientOptions.ServiceVersion.V2019_05_06_Preview)]
+    [ClientTestFixture(SearchClientOptions.ServiceVersion.V2020_06_30)]
     public abstract partial class SearchTestBase : RecordedTestBase<SearchTestEnvironment>
     {
         /// <summary>